wgpu: Split off code that actually renders a frame
